[INTERNAL] i2c: xlp9xx: Check for Bus state after every transfer
I2C bus enters the STOP condition after the DATA_DONE interrupt is raised.
Essentially the driver should be checking the bus state before sending
the next transaction. In case the next transaction is initiated while the
bus is busy, the prior transactions stop condition is not acheived.
Add the check to make sure the bus is not busy before next transaction.

Signed-off-by: George Cherian <george.cherian@cavium.com>
Signed-off-by: Vadim Lomovtsev <Vadim.Lomovtsev@cavium.com>