Fix doorbell bar mapping for dpi > 1
The current implementation checked that a doorbell address is in a db_size range always from first db physical address. This means that trying to map doorbells to a different dpi address on the bar resulted in the mapping being done without the write_combine option. This lead to doorbells not occurring on some system and left application to hang